Redis Create a Redic client:
Create a function reserveSeat, that will take into argument number, and set the key available_seats with the number
Create a function getCurrentAvailableSeats, it will return the current number of available seats (by using promisify for Redis)
When launching the application, set the number of available to 50
Initialize the boolean reservationEnabled to true - it will be turn to false when no seat will be available
Kue queue
Create a Kue queue

Server
Create an express server listening on the port 1245. (You will start it via: npm run dev 100-seat.js).

Requirements:

Make sure to use promisify with Redis
Make sure to use the await/async keyword to get the value from Redis
Make sure the format returned by the web application is always JSON and not text
Make sure that only the allowed amount of seats can be reserved
Make sure that the main route is displaying the right number of seats